requirements.txt
azure-cognitiveservices-vision-computervision
msrest
azure-ai-formrecognizer